Move the SessionCreationSteps into a separate file in spec/support
